Changed the datatype of number to long

+++ phonebook.cpp
@@ -11,13 +11,13 @@ namespace py = pybind11;
 
 class Phonebook {
     private:
-        std::map<std::string, int> contact;
+        std::map<std::string, long long> contact;
     public:
         Phonebook() = default;
-        Phonebook(const std::map<std::string, int> &initial) {
+        Phonebook(const std::map<std::string, long long> &initial) {
             contact = initial;
         }
-        void add_contact(const std::string &name, const int &phone_no) {
+        void add_contact(const std::string &name, const long long &phone_no) {
             contact[name] = phone_no;
         }
         bool remove_contact(const std::string &name) {
@@ -36,7 +36,7 @@ class Phonebook {
             }
             std::cout << "Not found\n";
         }
-        std::map<std::string, int> get_AllContacts() const {
+        std::map<std::string, long long> get_AllContacts() const {
             return contact;
         }
 };
@@ -45,9 +45,10 @@ PYBIND11_MODULE(phonebook,m) {
     m.doc() = "Phonebook with cpp as backend";
     py::class_<Phonebook>(m, "Phonebook")
         .def(py::init<>())   //default constructor
-        .def(py::init<const std::map<std::string, int>&>())
+        .def(py::init<const std::map<std::string, long long>&>())
         .def("add_contact", &Phonebook::add_contact)
         .def("remove_contact", &Phonebook::remove_contact)
         .def("get_number", &Phonebook::get_number)
         .def("get_AllContacts", &Phonebook::get_AllContacts);
+
 }
